+++ utils/catalog-parser/src/test/java/de/sovity/edc/utils/catalog/DspCatalogServiceTest.java
@@ -45,7 +45,8 @@ class DspCatalogServiceTest {
 
         var result = CompletableFuture.completedFuture(StatusResult.success(catalogJson.getBytes(StandardCharsets.UTF_8)));
         when(catalogService.requestCatalog(eq(endpoint), eq("dataspace-protocol-http"), eq(QuerySpec.max()))).thenReturn(result);
-        var dataOfferBuilder = new DspDataOfferBuilder(new TitaniumJsonLd(mock(Monitor.class)));
+        var monitor = mock(Monitor.class);
+        var dataOfferBuilder = new DspDataOfferBuilder(new TitaniumJsonLd(monitor));
 
         return new DspCatalogService(catalogService, dataOfferBuilder);
     }
@@ -71,7 +72,7 @@ class DspCatalogServiceTest {
 
         assertThat(offer.getContractOffers()).hasSize(1);
         var co = offer.getContractOffers().get(0);
-        assertThat(co.getContractOfferId()).isEqualTo("policy-1");
+        assertThat(co.getContractOfferId()).isEqualTo("contract-id:asset-id:ODJjMDNjMmM4YzQ5NmE0OTg4ZjVjOTY4OGU2MmMyN2UxYjIxZDAwZQ==");
         assertThat(toJson(co.getPolicyJsonLd())).contains("ALWAYS_TRUE");
 
         assertThat(offer.getDistributions()).hasSize(1);